Fix architecture and naming standards
Moved UI States to presentation layer, standardized package naming to lowercase (selectchart/createschart), and added KDocs. Renamed mock data to VisualizationMockData and restored original developer comments.
