+++ app/src/main/java/com/oracle/visualize/presentation/screens/selectchart/ChartSelectionViewModel.kt
@@ -1,28 +1,43 @@
-package com.oracle.visualize.presentation.screens.ChartSelection
+package com.oracle.visualize.presentation.screens.selectchart
 
 import androidx.lifecycle.ViewModel
-import com.oracle.visualize.domain.models.ChartSelectionUiState
-import com.oracle.visualize.domain.models.MockData
-import com.oracle.visualize.domain.models.VisualizationSelection
+import com.oracle.visualize.domain.models.VisualizationMockData
 import kotlinx.coroutines.flow.MutableStateFlow
 import kotlinx.coroutines.flow.StateFlow
 import kotlinx.coroutines.flow.asStateFlow
 import kotlinx.coroutines.flow.update
 
+/**
+ * ViewModel for the Chart Selection screen.
+ * Handles the logic for selecting visualizations and editing their titles.
+ * Aligned with MVVM and Clean Architecture.
+ */
 class ChartSelectionViewModel : ViewModel() {
 
     private val _uiState = MutableStateFlow<ChartSelectionUiState>(ChartSelectionUiState.Loading)
+    
+    /**
+     * The UI state of the screen, exposed as a [StateFlow].
+     */
     val uiState: StateFlow<ChartSelectionUiState> = _uiState.asStateFlow()
 
     init {
         loadMockCharts()
     }
 
+    /**
+     * Loads the mock visualizations from the domain layer provider [VisualizationMockData].
+     */
     private fun loadMockCharts() {
-        val mockCharts = MockData.visualizations.map { VisualizationSelection(it) }
+        // We obtain the mock data from our specialized domain model provider
+        val mockCharts = VisualizationMockData.visualizations.map { VisualizationSelection(it) }
         _uiState.value = ChartSelectionUiState.Success(charts = mockCharts)
     }
 
+    /**
+     * Toggles the selection state of a specific visualization.
+     * @param chartId The unique identifier of the visualization.
+     */
     fun toggleSelection(chartId: String) {
         val currentState = _uiState.value
         if (currentState is ChartSelectionUiState.Success) {
@@ -37,6 +52,11 @@ class ChartSelectionViewModel : ViewModel() {
         }
     }
 
+    /**
+     * Updates the title of a specific visualization.
+     * @param chartId The unique identifier of the visualization.
+     * @param newTitle The new title string.
+     */
     fun updateChartTitle(chartId: String, newTitle: String) {
         val currentState = _uiState.value
         if (currentState is ChartSelectionUiState.Success) {
@@ -52,6 +72,10 @@ class ChartSelectionViewModel : ViewModel() {
         }
     }
 
+    /**
+     * Toggles the visibility of the unsaved changes dialog.
+     * @param show True to show the dialog, false to hide it.
+     */
     fun showUnsavedChangesDialog(show: Boolean) {
         val currentState = _uiState.value
         if (currentState is ChartSelectionUiState.Success) {
@@ -59,6 +83,10 @@ class ChartSelectionViewModel : ViewModel() {
         }
     }
 
+    /**
+     * Checks if at least one visualization is selected.
+     * @return True if any chart is selected, false otherwise.
+     */
     fun hasSelections(): Boolean {
         val currentState = _uiState.value
         return if (currentState is ChartSelectionUiState.Success) {
